feat: implement authentication pages and email functionality
