+++ .env.example
@@ -1,6 +1,13 @@
+# NEXT AUTH
 AUTH_SECRET="change-me"
 
-NEXT_PUBLIC_COMPANY_NAME=''
+# GITHUB
 GH_TOKEN='<YOUR_GITHUB_TOKEN>'
 DATA_REPOSITORY='<DATA_REPOSITORY>'
+
+# DATABASE
 DATABASE_URL=
+
+# API KEYS
+RESEND_API_KEY=
+NOVU_API_KEY=
