+++ app/[locale]/auth/forgot-password/page.tsx
@@ -1,0 +1,3 @@
+export default function ForgotPasswordPage() {
+  return null;
+}
