+++ app/[locale]/auth/register/page.tsx
@@ -1,4 +1,4 @@
-import { AuthPage } from "../login-form";
+import { AuthPage } from "../components/login-form";
 
 export default function LoginPage() {
   return <AuthPage form="signup" />;
