+++ lib/db/schema.ts
@@ -105,6 +105,13 @@ export const activityLogs = pgTable("activityLogs", {
   ipAddress: varchar("ipAddress", { length: 45 }),
 });
 
+export const passwordResetTokens = pgTable("passwordResetTokens", {
+  id: serial("id").primaryKey(),
+  email: text("email").notNull(),
+  token: text("token").notNull().unique(),
+  expires: timestamp("expires", { mode: "date" }).notNull(),
+});
+
 export type NewUser = typeof users.$inferInsert;
 export type ActivityLog = typeof activityLogs.$inferSelect;
 export type NewActivityLog = typeof activityLogs.$inferInsert;
