+++ lib/mail.ts
@@ -1,0 +1,37 @@
+import { Resend } from "resend";
+import { getCachedConfig } from "./content";
+
+const resend = new Resend(process.env.RESEND_API_KEY);
+
+export const sendVerificationEmail = async (email: string, token: string) => {
+  const config = await getCachedConfig();
+  const confirmLink = `${config?.app_url}/auth/new-verification?token=${token}`;
+
+  await resend.emails.send({
+    from: "onboarding@resend.dev",
+    to: email,
+    subject: "Confirm your email",
+    html: `<p>Click <a href="${confirmLink}">here</a> to confirm email.</p>`,
+  });
+};
+
+export const sendPasswordResetEmail = async (email: string, token: string) => {
+  const config = await getCachedConfig();
+  const resetLink = `${config?.app_url}/auth/new-password?token=${token}`;
+
+  await resend.emails.send({
+    from: "onboarding@resend.dev",
+    to: email,
+    subject: "Reset your password",
+    html: `<p>Click <a href="${resetLink}">here</a> to reset password.</p>`,
+  });
+};
+
+export const sendTwoFactorTokenEmail = async (email: string, token: string) => {
+  await resend.emails.send({
+    from: "onboarding@resend.dev",
+    to: email,
+    subject: "2FA Code",
+    html: `<p>Your 2FA code: ${token}</p>`,
+  });
+};
