+++ messages/es.json
@@ -4,7 +4,8 @@
     "ABOUT": "Acerca de",
     "DISCOVER": "Descubrir",
     "LOGIN": "Acceso",
-    "SIGN_UP": "Inscribirse"
+    "SIGN_UP": "Inscribirse",
+    "OR_CONTINUE_WITH": "O continuar con"
   },
   "footer": {
     "ALL_RIGHTS_RESERVED": "Reservados todos los derechos"
@@ -14,5 +15,21 @@
     "BEST_DIRECTORY_WEBSITE_TEMPLATE": "La mejor plantilla de sitio web de directorio",
     "CATEGORIES": "Categorías",
     "THIS_IS_A_DEMO_DIRECTORY_WEBSITE": "Este es un sitio web de directorio de demostración."
+  },
+  "auth": {
+    "ALREADY_HAVE_AN_ACCOUNT": "¿Ya tienes una cuenta? \nIniciar sesión",
+    "CREATE_ACCOUNT": "Crear una cuenta",
+    "EMAIL_ADDRESS": "Dirección de correo electrónico",
+    "ENTER_YOUR_CREDENTIALS_HEADER": "Ingrese sus credenciales para acceder a su cuenta",
+    "ENTER_YOUR_EMAIL": "Ingrese su correo electrónico",
+    "ENTER_YOUR_FULL_NAME": "Ingrese su nombre completo",
+    "ENTER_YOUR_PASSWORD": "Ingrese su contraseña",
+    "FILL_IN_OUR_DIRECTORY": "Complete el formulario para unirse a nuestro directorio",
+    "FORGOT_PASSWORD": "¿Has olvidado tu contraseña?",
+    "FULL_NAME": "Nombre completo",
+    "NEED_AN_ACCOUNT": "¿Necesitas una cuenta? \nInscribirse",
+    "PASSWORD": "Contraseña",
+    "REMEMBER_ME": "Acuérdate de mí",
+    "SIGN_IN": "Iniciar sesión"
   }
 }
